+++ samples/hello/hello_world.rb
@@ -1,0 +1,7 @@
+require 'glimmer-dsl-swing'
+
+include Glimmer
+
+jframe('Hello, World!') {
+  jlabel('Hello, World!')
+}.show