+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 ## 0.0.1
 
 - Initial version of Glimmer DSL for Swing
+- General component support
+- General component property support
+- Hello, World! sample